refactor folder organization by component
