+++ src/components/recentTrades/duck.js
