+++ src/components/trade/duck.js
