+++ src/components/tradeHelper/duck.js
