+++ src/components/tradeHelper/tradeHelper.jsx
