+++ src/components/tradingActivity/tradingActivity.jsx
